remove separate header declaring forward declarations.
